# [terra-brand-footer] Migrate terra-brand-footer, optimize babel compile.