--- a/package.json
+++ b/package.json
@@ -42,7 +42,7 @@
     "clean:install": "npm run clean && npm install",
     "clean:obsolete-snapshots": "npm test -- -u",
     "compile": "npm run compile:clean && npm run compile:build",
-    "compile:build": "lerna ls -a | sed 's/@cerner\\///g' | cut -d ' ' -f1 | xargs -I {} babel packages/{}/src --out-dir packages/{}/lib --copy-files",
+    "compile:build": "lerna run compile --parallel",
     "compile:clean": "rm -rf packages/**/lib",
     "compile:src": "lerna run compile",
     "compile:prod": "rm -rf ./build; NODE_ENV=production webpack --mode=production",