--- a/packages/terra-brand-footer/CHANGELOG.md
+++ b/packages/terra-brand-footer/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Migrated doc to terra-framework-docs.
+
 ## 3.8.0 - (September 20, 2021)
 
 * Changed