--- a/packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/BrandFooter.1.doc.mdx
+++ b/packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/BrandFooter.1.doc.mdx
@@ -2,8 +2,8 @@ import { Badge } from 'terra-brand-footer/package.json?dev-site-package';
 
 import BrandFooterPropsTable from 'terra-brand-footer/src/BrandFooter?dev-site-props-table';
 
-import DefaultBrandFooter from '../example/DefaultBrandFooter?dev-site-example';
-import VerticalBrandFooter from '../example/VerticalBrandFooter?dev-site-example';
+import DefaultBrandFooter from './example/DefaultBrandFooter?dev-site-example';
+import VerticalBrandFooter from './example/VerticalBrandFooter?dev-site-example';
 
 <Badge />