--- a/packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/UpgradeGuide.2.doc.mdx
+++ b/packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/UpgradeGuide.2.doc.mdx
@@ -11,7 +11,7 @@ import { Badge } from 'terra-brand-footer/package.json?dev-site-package';
 ### Uplift to V3
 ```diff
 import React from 'react';
-import BrandFooter from '../../../BrandFooter';
+import BrandFooter from 'terra-brand-footer';
 
 class Default extends React.Component {
   constructor() {