--- a/packages/terra-framework-docs/src/terra-dev-site/test/brand-footer/ManySectionsBrandFooter.test.jsx
+++ b/packages/terra-framework-docs/src/terra-dev-site/test/brand-footer/ManySectionsBrandFooter.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import BrandFooter from '../../../../../terra-brand-footer/src/BrandFooter';
+import BrandFooter from 'terra-brand-footer';
 
 export default () => (
   <BrandFooter