--- a/packages/terra-framework-docs/src/terra-dev-site/test/brand-footer/NoNavBrandFooter.test.jsx
+++ b/packages/terra-framework-docs/src/terra-dev-site/test/brand-footer/NoNavBrandFooter.test.jsx
@@ -1,29 +1,13 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 
-import BrandFooter from '../../../BrandFooter';
+import BrandFooter from 'terra-brand-footer';
 import styles from './BrandFooterTestCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
 export default () => (
   <BrandFooter
-    isVertical={false}
-    sections={[
-      {
-        headerText: 'Links',
-        links: [
-          { text: 'Terra UI', href: 'http://terra-ui.com/static/#/site/home' },
-          { text: 'Cerner', href: 'https://www.cerner.com/', target: '_blank' },
-        ],
-      },
-      {
-        headerText: 'More Links',
-        links: [
-          { text: 'Cerner Engineering', href: 'https://engineering.cerner.com/' },
-        ],
-      },
-    ]}
     contentLeft={(
       <a
         href="http://terra-ui.com/"