… packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/ChangeLog.3.doc.mdx
… packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/example/BrandFooterExampleCommon.module.scss
… packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/example/DefaultBrandFooter.jsx
… packages/terra-framework-docs/src/terra-dev-site/doc/brand-footer/example/VerticalBrandFooter.jsx
… packages/terra-framework-docs/src/terra-dev-site/test/brand-footer/BrandFooterTestCommon.module.scss