### Summary
<!--- Summarize and explain the reason behind these code changes. What are the changes, and why are they necessary? -->

Moving terra-brand-footer is pretty straight forward.

npm compile:build numbers: 

Before:
npm run compile:build  83.62s user 9.58s system 132% cpu 1:10.28 total

After:
npm run compile:build  120.05s user 17.39s system 417% cpu 32.925 total


<!--- Include any issue addressed by this pull request. -->
<!--- Example: Closes #45 -->
Closes #

### Deployment Link
<!---Include the deployment link, if applicable. -->
<!--- Example: https://terra-framework-deployed-pr-45.herokuapp.com/ -->
https://terra-framework-deployed-pr-#.herokuapp.com/

### Testing
<!-- Demonstrate that these changes are stable. How have these changes been verified? -->

### Additional Details
<!-- List anything else that is relevant to this issue. Additional information will help us better understand your changes and speed up the review process. -->

<!--
*Before publishing*

1. Assign yourself to the PR.
2. Add the appropriate labels
3. Add your name to the CONTRIBUTORS.md file. Adding your name to the CONTRIBUTORS.md file signifies agreement to all rights and reservations provided by the License.
-->

Thank you for contributing to Terra.
@cerner/terra